+++ src/core/api/auth/controller/auth.controller.js
@@ -1,12 +1,16 @@
 import { AuthService } from '../../../modules/auth/service/auth.service';
 import { LoginDto } from '../../../modules/auth/dto/index';
+import { ValidHttpResponse } from '../../../../packages/handler/response/validHttp.response';
 
 class Controller {
     constructor() {
         this.service = AuthService;
     }
 
-    login = req => this.service.login(LoginDto(req.body))
+    login = async req => {
+        const data = await this.service.login(LoginDto(req.body));
+        return ValidHttpResponse.toOkResponse(data);
+    }
 }
 
 export const AuthController = new Controller();