+++ src/core/api/user/controller/user.controller.js
@@ -3,6 +3,7 @@ import { RequestTransformer } from '../../../../packages/restBuilder/core/reques
 import SearchUserSchema from '../query/searchUser.schema.json';
 import { Pageable, PageableMeta } from '../../../../packages/restBuilder/core/pageable';
 import { CreateUserDto } from '../../../modules/user/dto/createUser.dto';
+import { ValidHttpResponse } from '../../../../packages/handler/response/validHttp.response';
 
 class Controller {
     constructor() {
@@ -12,25 +13,37 @@ class Controller {
     findAll = async req => {
         const reqTransformed = new RequestTransformer(req.query, SearchUserSchema);
         const data = await this.service.findAll(reqTransformed.translate());
-        const count = await this.service.count();
-        return Pageable.of(data)
+        const pagedData = Pageable.of(data[0])
             .addMeta(
                 PageableMeta
                     .builder()
                     .appendRequestFormation(reqTransformed)
-                    .appendTotalRecord(count)
+                    .appendTotalRecord(data[1])
                     .build()
             )
             .build();
+        return ValidHttpResponse.toOkResponse(pagedData);
     }
 
-    createOne = req => this.service.createOne(CreateUserDto(req.body))
+    createOne = async req => {
+        const data = await this.service.createOne(CreateUserDto(req.body));
+        return ValidHttpResponse.toCreatedResponse(data);
+    }
 
-    findOne = req => this.service.findOne(req.params)
+    findOne = async req => {
+        const data = await this.service.findOne(req.params);
+        return ValidHttpResponse.toOkResponse(data);
+    }
 
-    patchOne = req => this.service.patchOne(req.params, req.body)
+    patchOne = async req => {
+        await this.service.patchOne(req.params, req.body);
+        return ValidHttpResponse.toNoContentResponse();
+    }
 
-    deleteOne = req => this.service.deleteOne(req.params)
+    deleteOne = async req => {
+        await this.service.deleteOne(req.params);
+        return ValidHttpResponse.toNoContentResponse();
+    }
 }
 
 export const UserController = new Controller();